lock: change locking to a mutex
Mutexes have the advantage that they allow for blocking waits natively.
These are typically handled by some external sync mechanism.
Thusly, threads will be put to sleep by the OS until the other place has
unlocked it.
The old atomic-bool solution required a busy-wait to achieve this.

This changed implementation has the disadvantage that it requires rustc
version >= 1.70.0 which, as of commiting this, still is in beta.
However OnceLock avoids having to pull in once_cell or lazy_static as
dependencies. So i think its best to wait until june and then merge
this.